Corrected a mistake in the product reviews view

+++ resources/views/pages/product.blade.php
@@ -73,10 +73,12 @@
                                     <div id="review{{ $review->idusers }}_{{ $review->idproduct }}">
                                         <div class="review_header">
                                             <p style="font-weight: normal; font-size: 18px;"><strong>{{ $review->name }}</strong>, {{ $review->reviewdate }}</p>
-                                            <div class="header_buttons" id="review_buttons{{ $review->idusers }}_{{ $review->idproduct }}">
-                                                <button onclick="editReview({{ $review->idusers }}, {{ $review->idproduct }})" style="all: unset; margin-right: 8px; cursor: pointer;"><i class='fas fa-pencil-alt' style='font-size: 24px'></i></button>
-                                                <button onclick="deleteReview({{ $review->idusers }}, {{ $review->idproduct }})" style="all: unset; cursor: pointer;"><i class="fa fa-trash" aria-hidden="true" style='font-size: 24px'></i></button>
-                                            </div>
+                                            @if(Auth::check() && Auth::user()->id == $review->idusers)
+                                                <div class="header_buttons" id="review_buttons{{ $review->idusers }}_{{ $review->idproduct }}">
+                                                    <button onclick="editReview({{ $review->idusers }}, {{ $review->idproduct }})" style="all: unset; margin-right: 8px; cursor: pointer;"><i class='fas fa-pencil-alt' style='font-size: 24px'></i></button>
+                                                    <button onclick="deleteReview({{ $review->idusers }}, {{ $review->idproduct }})" style="all: unset; cursor: pointer;"><i class="fa fa-trash" aria-hidden="true" style='font-size: 24px'></i></button>
+                                                </div>
+                                            @endif
                                         </div>
                                         <p style="font-weight:normal" id="review_content{{ $review->idusers }}_{{ $review->idproduct }}">{{ $review->content }}</p>
                                         <div class="ratings">
